moved parts of map to modules. We will see if it is a good idea.
